VPN Label

+++ Server/src/bgp/bgp_common.h
@@ -132,7 +132,7 @@ namespace bgp {
          std::string    rd_administrator_subfield;
          std::string    rd_assigned_number;
          uint8_t        rd_type;
-         uint8_t        vpn_label;
+         uint32_t       vpn_label;
      };
 
     /*********************************************************************//**
